+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
+to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1] - unreleased
+## [0.0.2] - unreleased
+
+### Added
+
+* Added `packet.IsLikeModbusTCP()` to check if given bytes are possibly TCP packet or start of packet.
+* Added `Parse*Request*` for every function type to help implement Modbus servers.
+
+## [0.0.1] - 2021-04-11
 
 ### Added